+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoPlayer.kt
@@ -76,6 +76,7 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
       playbackServiceBinder?.service?.setShowNotification(value, this.player)
     }
   var duration = 0f
+  var isLive = false
 
   private var serviceConnection: ServiceConnection
   internal var playbackServiceBinder: PlaybackServiceBinder? = null
@@ -125,6 +126,7 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
       val newVideoSource = VideoManager.getVideoSourceFromMediaItem(mediaItem)
       this@VideoPlayer.videoSource = newVideoSource
       this@VideoPlayer.duration = 0f
+      this@VideoPlayer.isLive = false
       if (reason == Player.MEDIA_ITEM_TRANSITION_REASON_REPEAT) {
         sendEventOnJSThread("playToEnd")
       }
@@ -137,6 +139,7 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
       }
       if (playbackState == Player.STATE_READY) {
         this@VideoPlayer.duration = this@VideoPlayer.player.duration / 1000f
+        this@VideoPlayer.isLive = this@VideoPlayer.player.isCurrentMediaItemLive
       }
       setStatus(playerStateToPlayerStatus(playbackState), null)
       super.onPlaybackStateChanged(playbackState)
@@ -156,6 +159,7 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
       error?.let {
         setStatus(ERROR, error)
         this@VideoPlayer.duration = 0f
+        this@VideoPlayer.isLive = false
       } ?: run {
         setStatus(playerStateToPlayerStatus(player.playbackState), null)
       }